Rename addonSlice to searchSlice
